Break density button into a separate component
