+++ react/components/TableCompose/Toolbar.js
@@ -7,19 +7,18 @@ import ButtonWithIcon from '../ButtonWithIcon'
 import ButtonGroup from '../ButtonGroup'
 import Toggle from '../Toggle'
 import IconColumns from '../icon/Columns'
-import IconDensity from '../icon/Density'
 import IconDownload from '../icon/Download'
 import IconPlus from '../icon/Plus'
 import IconUpload from '../icon/Upload'
 import IconOptionsDots from '../icon/OptionsDots'
 import useOutsideClick from './useOutsideCick'
 import ToolbarInput from './ToolbarInput'
+import DensityBtn from './DensityBtn'
 
 const MAX_FIELDS_BOX_HEIGHT = 192
 const FIELDS_BOX_ITEM_HEIGHT = 36
 const FIELDS_BOX_WIDTH = 292
 const BOX_SHADOW_STYLE = { boxShadow: '0px 1px 18px rgba(0, 0, 0, 0.14)' }
-const DENSITY_OPTIONS = ['low', 'medium', 'high']
 const ICON_OPTICAL_COMPENSATION = { marginTop: '1.5px' }
 const LIGHT_ICON_SIZE = 16
 const MEDIUM_ICON_SIZE = 14
@@ -46,20 +45,12 @@ const Toolbar = ({
   loading,
 }) => {
   const [isFieldsBoxVisible, setFieldsBoxVisible] = useState(false)
-  const [isDensityBoxVisible, setDensityBoxVisible] = useState(false)
   const fieldsBtnRef = useRef(null)
-  const densityBtnRef = useRef(null)
 
   useOutsideClick(
     fieldsBtnRef,
     () => setFieldsBoxVisible(false),
     isFieldsBoxVisible
-  )
-
-  useOutsideClick(
-    densityBtnRef,
-    () => setDensityBoxVisible(false),
-    isDensityBoxVisible
   )
 
   const calculateFieldsBoxHeight = () => {
@@ -108,62 +99,12 @@ const Toolbar = ({
       )}
       <div className="flex flex-row items-center">
         {isDensityVisible && (
-          <div
-            id="toggleDensity"
-            title={density.buttonLabel}
-            ref={densityBtnRef}
-            className="relative mh2">
-            <ButtonWithIcon
-              icon={
-                <span
-                  className="c-on-base mh2"
-                  style={ICON_OPTICAL_COMPENSATION}>
-                  <IconDensity size={MEDIUM_ICON_SIZE} />
-                </span>
-              }
-              disabled={loading}
-              variation="tertiary"
-              size="small"
-              // TODO: REVIEW
-              onClick={() => setDensityBoxVisible(!isDensityBoxVisible)}
-            />
-            {isDensityBoxVisible && (
-              <div
-                className={`absolute ${
-                  density.alignMenu === 'right' ? 'right-0' : 'left-0'
-                } z-999 ba b--muted-4 br2 mt2 mh2`}
-                style={BOX_SHADOW_STYLE}>
-                <div className="w-100 b2 br2 bg-base">
-                  <div
-                    style={{ height: 3 * FIELDS_BOX_ITEM_HEIGHT }}
-                    className="overflow-auto">
-                    {DENSITY_OPTIONS.map((key, index) => {
-                      const isKeySelected = selectedDensity === key
-                      return (
-                        <div
-                          key={index}
-                          className={`flex justify-between ph6 pv3 ${
-                            isKeySelected ? 'b--emphasis' : 'b--transparent'
-                          } pointer hover-bg-muted-5 bl bw1`}
-                          // TODO: REVIEW
-                          onClick={() => {
-                            handleToggleDensity(key)
-                            setDensityBoxVisible(false)
-                            density.handleCallback &&
-                              density.handleCallback(key)
-                          }}>
-                          <span
-                            className={`w-100 ${isKeySelected ? 'fw5' : ''}`}>
-                            {density[`${key}OptionLabel`]}
-                          </span>
-                        </div>
-                      )
-                    })}
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
+          <DensityBtn
+            density={density}
+            disabled={loading}
+            handleToggleDensity={handleToggleDensity}
+            selectedDensity={selectedDensity}
+          />
         )}
         {isFieldsVisible && (
           <div
